requirements.txt
aniso8601==7.0.0
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.1
pkg-resources==0.0.0
promise==2.3
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
SQLAlchemy==1.3.12
